+++ sample/source/conf.py
@@ -1,12 +1,8 @@
 """
 Sample ``conf.py``.
 """
-import os
-import sys
-sys.path.insert(0, os.path.abspath('../../src'))
 
 extensions = [
-    'sphinx-prompt',
     'sphinx_substitution_extensions',
 ]