+++ src/sphinx_substitution_extensions/utils.py
@@ -1,8 +1,0 @@
-def exists_dependency(
-    name: str,
-) -> bool:
-    try:
-        __import__(name)
-    except ImportError as e:
-        return False
-    return True