+++ tests/test_substitution_extensions_extras.py
@@ -1,19 +1,24 @@
-"""
-Tests for Sphinx extensions.
-"""
-
 import subprocess
 import sys
 from pathlib import Path
 from textwrap import dedent
 
+import pytest
+
 from sphinx_substitution_extensions import _exists_dependency
 
+_EXISTS_PROMPT_EXTENSION = _exists_dependency('sphinx-prompt')
+_REASON = 'requires sphinx-prompt to be installed'
 
-def test_no_substitution_code_block(tmp_path: Path) -> None:
+
+@pytest.mark.skipif(
+    not _EXISTS_PROMPT_EXTENSION,
+    reason=_REASON,
+)
+def test_prompt_specified_late(tmp_path: Path) -> None:
     """
-    The ``code-block`` directive does not replace the placeholders defined in
-    ``conf.py`` when not specified.
+    If sphinx-prompt is not specified in extensions before Sphinx substitution
+    extensions, an warning is given.
     """
     source_directory = tmp_path / 'source'
     source_directory.mkdir()
@@ -23,21 +28,10 @@ def test_no_substitution_code_block(tmp_path: Path) -> None:
     source_file.touch()
     conf_py_content = dedent(
         """\
-        extensions = ['sphinx_substitution_extensions']
-        rst_prolog = '''
-        .. |a| replace:: example_substitution
-        '''
+        extensions = ['sphinx_substitution_extensions', 'sphinx-prompt']
         """,
     )
     conf_py.write_text(conf_py_content)
-    source_file_content = dedent(
-        """\
-        .. code-block:: bash
-
-           $ PRE-|a|-POST
-        """,
-    )
-    source_file.write_text(source_file_content)
     destination_directory = tmp_path / 'destination'
     args = [
         sys.executable,
@@ -53,16 +47,29 @@ def test_no_substitution_code_block(tmp_path: Path) -> None:
         # Source file to process.
         str(source_file),
     ]
-    subprocess.check_output(args=args)
-    expected = 'PRE-example_substitution-POST'
-    content_html = Path(str(destination_directory)) / 'index.html'
-    assert expected not in content_html.read_text()
+    result = subprocess.run(
+        args=args,
+        check=False,
+        stderr=subprocess.PIPE,
+    )
+
+    expected_message = (
+        'sphinx-prompt must be in the conf.py extensions list before '
+        'sphinx_substitution_extensions'
+    )
+
+    assert result.returncode == 0  # Do not raise an error
+    assert expected_message in result.stderr.decode()
 
 
-def test_substitution_code_block(tmp_path: Path) -> None:
+@pytest.mark.skipif(
+    not _EXISTS_PROMPT_EXTENSION,
+    reason=_REASON,
+)
+def test_prompt_not_specified(tmp_path: Path) -> None:
     """
-    The ``code-block`` directive replaces the placeholders defined in
-    ``conf.py`` as specified.
+    If sphinx-prompt is not specified in extensions but is installed,
+    a warning is given.
     """
     source_directory = tmp_path / 'source'
     source_directory.mkdir()
@@ -73,6 +80,57 @@ def test_substitution_code_block(tmp_path: Path) -> None:
     conf_py_content = dedent(
         """\
         extensions = ['sphinx_substitution_extensions']
+        """,
+    )
+    conf_py.write_text(conf_py_content)
+    destination_directory = tmp_path / 'destination'
+    args = [
+        sys.executable,
+        '-m',
+        'sphinx',
+        '-b',
+        'html',
+        '-W',
+        # Directory containing source and configuration files.
+        str(source_directory),
+        # Directory containing build files.
+        str(destination_directory),
+        # Source file to process.
+        str(source_file),
+    ]
+    result = subprocess.run(
+        args=args,
+        check=False,
+        stderr=subprocess.PIPE,
+    )
+
+    expected_message = (
+        'sphinx-prompt must be in the conf.py extensions list before '
+        'sphinx_substitution_extensions'
+    )
+
+    assert result.returncode == 0  # Do not raise an error
+    assert expected_message in result.stderr.decode()
+
+
+@pytest.mark.skipif(
+    not _EXISTS_PROMPT_EXTENSION,
+    reason=_REASON,
+)
+def test_substitution_prompt(tmp_path: Path) -> None:
+    """
+    The ``prompt`` directive replaces the placeholders defined in ``conf.py``
+    when requested.
+    """
+    source_directory = tmp_path / 'source'
+    source_directory.mkdir()
+    source_file = source_directory / 'index.rst'
+    conf_py = source_directory / 'conf.py'
+    conf_py.touch()
+    source_file.touch()
+    conf_py_content = dedent(
+        """\
+        extensions = ['sphinx-prompt', 'sphinx_substitution_extensions']
         rst_prolog = '''
         .. |a| replace:: example_substitution
         '''
@@ -81,7 +139,7 @@ def test_substitution_code_block(tmp_path: Path) -> None:
     conf_py.write_text(conf_py_content)
     source_file_content = dedent(
         """\
-        .. code-block:: bash
+        .. prompt:: bash $
            :substitutions:
 
            $ PRE-|a|-POST
@@ -109,9 +167,13 @@ def test_substitution_code_block(tmp_path: Path) -> None:
     assert expected in content_html.read_text()
 
 
-def test_substitution_code_block_case_preserving(tmp_path: Path) -> None:
+@pytest.mark.skipif(
+    not _EXISTS_PROMPT_EXTENSION,
+    reason=_REASON,
+)
+def test_substitution_prompt_is_case_preserving(tmp_path: Path) -> None:
     """
-    The ``code-block`` directive respects the original case of replacements.
+    The ``prompt`` directive respects the original case of replacements.
     """
     source_directory = tmp_path / 'source'
     source_directory.mkdir()
@@ -121,7 +183,7 @@ def test_substitution_code_block_case_preserving(tmp_path: Path) -> None:
     source_file.touch()
     conf_py_content = dedent(
         """\
-        extensions = ['sphinx_substitution_extensions']
+        extensions = ['sphinx-prompt', 'sphinx_substitution_extensions']
         rst_prolog = '''
         .. |aBcD_eFgH| replace:: example_substitution
         '''
@@ -130,7 +192,7 @@ def test_substitution_code_block_case_preserving(tmp_path: Path) -> None:
     conf_py.write_text(conf_py_content)
     source_file_content = dedent(
         """\
-        .. code-block:: bash
+        .. prompt:: bash $
            :substitutions:
 
            $ PRE-|aBcD_eFgH|-POST
@@ -158,10 +220,14 @@ def test_substitution_code_block_case_preserving(tmp_path: Path) -> None:
     assert expected in content_html.read_text()
 
 
-def test_substitution_inline(tmp_path: Path) -> None:
+@pytest.mark.skipif(
+    not _EXISTS_PROMPT_EXTENSION,
+    reason=_REASON,
+)
+def test_no_substitution_prompt(tmp_path: Path) -> None:
     """
-    The ``substitution-code`` role replaces the placeholders defined in
-    ``conf.py`` as specified.
+    The ``prompt`` directive does not replace the placeholders defined in
+    ``conf.py`` when that is not requested.
     """
     source_directory = tmp_path / 'source'
     source_directory.mkdir()
@@ -171,7 +237,7 @@ def test_substitution_inline(tmp_path: Path) -> None:
     source_file.touch()
     conf_py_content = dedent(
         """\
-        extensions = ['sphinx_substitution_extensions']
+        extensions = ['sphinx-prompt', 'sphinx_substitution_extensions']
         rst_prolog = '''
         .. |a| replace:: example_substitution
         '''
@@ -180,7 +246,9 @@ def test_substitution_inline(tmp_path: Path) -> None:
     conf_py.write_text(conf_py_content)
     source_file_content = dedent(
         """\
-        Example :substitution-code:`PRE-|a|-POST`
+        .. prompt:: bash $
+
+           $ PRE-|a|-POST
         """,
     )
     source_file.write_text(source_file_content)
@@ -202,66 +270,4 @@ def test_substitution_inline(tmp_path: Path) -> None:
     subprocess.check_output(args=args)
     expected = 'PRE-example_substitution-POST'
     content_html = Path(str(destination_directory)) / 'index.html'
-    assert expected in content_html.read_text()
-
-
-def test_substitution_inline_case_preserving(tmp_path: Path) -> None:
-    """
-    The ``substitution-code`` role respects the original case of replacements.
-    """
-    source_directory = tmp_path / 'source'
-    source_directory.mkdir()
-    source_file = source_directory / 'index.rst'
-    conf_py = source_directory / 'conf.py'
-    conf_py.touch()
-    source_file.touch()
-    conf_py_content = dedent(
-        """\
-        extensions = ['sphinx_substitution_extensions']
-        rst_prolog = '''
-        .. |aBcD_eFgH| replace:: example_substitution
-        '''
-        """,
-    )
-    conf_py.write_text(conf_py_content)
-    source_file_content = dedent(
-        """\
-        Example :substitution-code:`PRE-|aBcD_eFgH|-POST`
-        """,
-    )
-    source_file.write_text(source_file_content)
-    destination_directory = tmp_path / 'destination'
-    args = [
-        sys.executable,
-        '-m',
-        'sphinx',
-        '-b',
-        'html',
-        '-W',
-        # Directory containing source and configuration files.
-        str(source_directory),
-        # Directory containing build files.
-        str(destination_directory),
-        # Source file to process.
-        str(source_file),
-    ]
-    subprocess.check_output(args=args)
-    expected = 'PRE-example_substitution-POST'
-    content_html = Path(str(destination_directory)) / 'index.html'
-    assert expected in content_html.read_text()
-
-
-def test_exists_dependency() -> None:
-    """
-    Test exist_dependency function.
-    """
-    dependency = 'sphinx_substitution_extensions'
-    assert _exists_dependency(dependency) is True
-
-
-def test_does_not_exists_dependency() -> None:
-    """
-    Test exist_dependency function.
-    """
-    dependency = 'fake_sphinx_substitution_extensions'
-    assert _exists_dependency(dependency) is False
+    assert expected not in content_html.read_text()